chore: partial implementation
