+++ src/orchestration/operators/batch/vep.py
@@ -25,70 +25,6 @@ if TYPE_CHECKING:
     from typing import Any
 
 
-class VepAnnotationPathManager:
-    """Manager class for setting correct mounting points for VEP google batch tasks."""
-
-    def __init__(
-        self,
-        vcf_input_path: str,
-        vep_output_path: str,
-        vep_cache_path: str,
-        mount_dir_root: str,
-    ):
-        self._mount_dir_root = mount_dir_root
-        self.paths = {
-            "input": GCSPath(vcf_input_path),
-            "output": GCSPath(vep_output_path),
-            "cache": GCSPath(vep_cache_path),
-        }
-
-    @cached_property
-    def mount_dir_root(self) -> str:
-        """Get the mount directory root."""
-        if not self._mount_dir_root.startswith("/"):
-            raise ValueError("Mount dir has to be an absolute path.")
-        if self._mount_dir_root.endswith("/"):
-            return str(Path(self._mount_dir_root))
-        return self._mount_dir_root
-
-    @cached_property
-    def path_registry(self) -> dict[str, GCSMountObject]:
-        """Get the path registry."""
-        return {
-            key: {
-                # NOTE: remote_path has to start from the bucket_name but without the gs://
-                # see https://cloud.google.com/batch/docs/create-run-job-storage#gcloud_2:~:text=BUCKET_PATH%3A%20the%20path,the%20subdirectory%20subdirectory.
-                "remote_path": f"{value.bucket}/{value.path}",
-                "mount_point": f"{self.mount_dir_root}/{key}",
-            }
-            for key, value in self.paths.items()
-        }
-
-    @cached_property
-    def cache_dir(self) -> str:
-        """Get cache dir."""
-        return self.path_registry["cache"]["mount_point"]
-
-    @cached_property
-    def input_dir(self) -> str:
-        """Get input dir."""
-        return self.path_registry["input"]["mount_point"]
-
-    @cached_property
-    def output_dir(self) -> str:
-        """Get output dir."""
-        return self.path_registry["output"]["mount_point"]
-
-    @cached_property
-    def mount_config(self) -> list[GCSMountObject]:
-        """Return the mount configuration.
-
-        Returns:
-            list[dict[str, str]]: The mount configuration.
-        """
-        return list(self.path_registry.values())
-
-
 class VepAnnotateOperator(GoogleCloudBaseOperator):
     """Annotate vcf files in batch job.
